+++ include/jary/modules.h
@@ -5,10 +5,9 @@
 
 int define_function(int			module,
 		    const char	       *key,
-		    size_t		keysz,
+		    uint32_t		keysz,
 		    struct jy_obj_func *func);
 
 const char *error_message(int status);
 
 #endif // JAYVM_MODULES_H
-
