+++ lib/jay/scanner.h
@@ -3,16 +3,15 @@
 
 #include "token.h"
 
-#include <stddef.h>
+#include <stdint.h>
 
 // line and ofs start from 1
 void jry_scan(const char  *src,
-	      size_t	   length,
+	      uint32_t	   length,
 	      enum jy_tkn *type,
-	      size_t	  *line,
-	      size_t	  *ofs,
+	      uint32_t	  *line,
+	      uint32_t	  *ofs,
 	      const char **lxstart,
 	      const char **lxend);
 
 #endif // JAYVM_SCAN_H
-
